+++ src/main/java/recommend_br_use_case/application_business/RecommendBROutputBoundary.java
@@ -1,4 +1,6 @@
-package recommend_br_use_case;
+package recommend_br_use_case.application_business;
+
+import recommend_br_use_case.application_business.RecommendBRResponseModel;
 
 /**
  * Interface used to do dependence inversion for RecommendBRPresenter.
