+++ src/main/java/recommend_br_use_case/application_business/RecommendBRResponseModel.java
@@ -1,4 +1,4 @@
-package recommend_br_use_case;
+package recommend_br_use_case.application_business;
 
 import retrieve_timetable_use_case.CourseModel;
