--- src/main/java/recommend_br_use_case/application_business/removeLater.java
@@ -1,5 +1,0 @@
-package recommend_br_use_case.application_business;
-
-public class removeLater {
-    // Remove this class later when you move in your classes
-}
