+++ src/main/java/recommend_br_use_case/frameworks_and_drivers/CourseInfoPanel.java
@@ -1,4 +1,4 @@
-package screens;
+package recommend_br_use_case.frameworks_and_drivers;
 
 import javax.swing.*;
