+++ src/main/java/recommend_br_use_case/frameworks_and_drivers/RecommendBRCourseViewModel.java
@@ -1,4 +1,4 @@
-package screens;
+package recommend_br_use_case.frameworks_and_drivers;
 
 import java.util.List;
