+++ src/main/java/recommend_br_use_case/frameworks_and_drivers/RecommendBROutputScreen.java
@@ -1,4 +1,4 @@
-package screens;
+package recommend_br_use_case.frameworks_and_drivers;
 
 import edit_timetable_use_case.EditTimetableController;
 import entities.InvalidSectionsException;
