+++ src/main/java/recommend_br_use_case/frameworks_and_drivers/RecommendBRWindow.java
@@ -1,6 +1,11 @@
-package screens;
+package recommend_br_use_case.frameworks_and_drivers;
 
 import edit_timetable_use_case.EditTimetableController;
+import recommend_br_use_case.frameworks_and_drivers.RecommendBRInputScreen;
+import recommend_br_use_case.frameworks_and_drivers.RecommendBROutputScreen;
+import recommend_br_use_case.frameworks_and_drivers.RecommendBRViewModel;
+import recommend_br_use_case.interface_adapters.IRecommendBRView;
+import recommend_br_use_case.interface_adapters.RecommendBRController;
 
 import javax.swing.*;
 import java.awt.event.WindowEvent;
