+++ src/main/java/recommend_br_use_case/interface_adapters/RecommendBRPresenter.java
@@ -1,9 +1,12 @@
-package screens;
+package recommend_br_use_case.interface_adapters;
 
-import recommend_br_use_case.*;
+import recommend_br_use_case.application_business.RecommendBROutputBoundary;
+import recommend_br_use_case.application_business.RecommendBRResponseModel;
+import recommend_br_use_case.frameworks_and_drivers.RecommendBRCourseViewModel;
 import retrieve_timetable_use_case.BlockModel;
 import retrieve_timetable_use_case.CourseModel;
 import retrieve_timetable_use_case.SectionModel;
+import recommend_br_use_case.frameworks_and_drivers.RecommendBRViewModel;
 
 import java.util.ArrayList;
 import java.util.List;
