--- src/main/java/recommend_br_use_case/interface_adapters/removeLater.java
@@ -1,5 +1,0 @@
-package recommend_br_use_case.interface_adapters;
-
-public class removeLater {
-    // Remove this class later when you move in your classes
-}
